Last controller implementation, postman examples, rm mapping from xml pojos

+++ src/main/java/no/regnskap/model/xml/RegnskapXmlPost.java
@@ -14,7 +14,7 @@ public class RegnskapXmlPost {
     @JacksonXmlProperty(isAttribute = true)
     private String nr;
 
-    private Long tall;
+    private BigDecimal tall;
     private String notehenvisning;
     private String fritekst;
 
@@ -34,12 +34,12 @@ public class RegnskapXmlPost {
         this.nr = nr;
     }
 
-    public Long getTall() {
+    public BigDecimal getTall() {
         return tall;
     }
 
     public void setTall(BigDecimal tall) {
-        this.tall = tall.longValue();
+        this.tall = tall;
     }
 
     public String getNotehenvisning() {
